Fixes to cross compile for windows 32/64 bits

+++ incl-master/incl-linux32/stddef.h
@@ -5,8 +5,8 @@
 #define __CF_STDDEF_H
 
 #if defined(_WIN32) && !defined(_WIN64)
-    typedef long ssize_t;
-    typedef unsigned long size_t;
+    typedef int ssize_t;
+    typedef unsigned int size_t;
 #elif defined(_WIN64)
     typedef long long ssize_t;
     typedef unsigned long long size_t;
